♻️ refactor : register 분리
